// journalist-client/src/model/bindings/GroupMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GroupId } from "./GroupId";
import type { GroupMessageContent } from "./GroupMessageContent";
import type { MessageId } from "./MessageId";
import type { SentinelIdentity } from "./SentinelIdentity";

/**
 * Represents a message that is stored in the vault and displayed in the UI,
 * including the content of the message and metadata like sender and timestamp.
 */
export type GroupMessage = {
  id: MessageId;
  sender: SentinelIdentity;
  group_id: GroupId;
  content: GroupMessageContent;
  read: boolean;
  published_at: string;
};